Add double type for money input

+++ database/migrations/2020_04_03_220657_create_tour_dates_table.php
@@ -17,7 +17,7 @@ class CreateTourDatesTable extends Migration
             $table->id();
             $table->date('date');
             $table->integer('seats');
-            $table->integer('price');
+            $table->double('price', 10, 2);
             $table->timestamps();
         });
     }
